refactor(provider): extract OpenRouter's model routing into its own class
The routing cluster — six methods, 163 contiguous lines — was the one part of
the adapter that is pure decision logic: no HTTP, no PSR-7, no response
objects. It is also the part most likely to grow, because routing strategies
are the product's premise. It now lives in Provider/OpenRouter/ModelRouter,
the first per-adapter collaborator; the pattern rules are recorded in ADR-125.

Bodies moved verbatim. The router is stateless: the strategy stays on the
provider, whose configure() owns and validates it, and arrives per call; the
catalogue arrives as a closure so a call that names its model explicitly
still performs no network fetch — exactly the short-circuit the inline code
had. The vision path keeps its eager fetch, as before.

The PHPat guard that keeps Service\* off concrete adapters matched only
classes ending in "Provider"; the new sub-namespace is added to its deny set
so a collaborator cannot quietly escape the rule the adapters live under.

Deliberately NOT extracted, per the seam analysis: the four request/response
methods (irreducible adapter surface), the streaming trio (already factored),
and sendOpenRouterRequest() — its divergences from AbstractProvider's
sendRequest() (no retries, 200-only, no audit-reason stamp) are load-bearing
behaviour that a "unification" would silently change.

All 130 provider tests pass unchanged; the ~22 routing tests assert through
the captured request body and never noticed the move.
